correct the interface
